+++ kaseev.andrey/P4/matrix.cpp
@@ -1,7 +1,8 @@
 #include "matrix.hpp"
 
 namespace MatrixStuff {
-  size_t NumberOfDiagonals(int *matrixPointer, int n, int m) {
+  size_t NumberOfDiagonals(int *matrixPointer, int n, int m)
+  {
     if (matrixPointer == nullptr)
     {
       return 0;
@@ -29,32 +30,6 @@ namespace MatrixStuff {
       {
         count++;
       }
-    }
-    return count;
-  }
-
-  void writeResult(std::ostream &output, int result, bool UpperTriangularMatrix)
-  {
-    if (UpperTriangularMatrix)
-    {
-      output << result << "\n" << "true";
-    }
-    else
-    {
-      output << result << "\n" << "false";
-    }
-  }
-
-  int readMatrix(std::istream &input, int *matrixPointer, int n, int m)
-  {
-    int count = 0;
-    for (int i = 0; i < n * m; ++i)
-    {
-      if (!(input >> matrixPointer[i]))
-      {
-        return count;
-      }
-      count += 1;
     }
     return count;
   }
